+++ NEWS.md
@@ -1,3 +1,13 @@
+# polAr 0.1.3.1
+
+**May 26, 2020: dev version**
+
+* Add first version of `map_results()` function to visualize election data in geographical format with polygons as geometry.
+
+* Minor fix in `geofacet` plots (`aspect.ratio` inside `ggplot2::theme()`).
+
+* Simplify visualization workflow adding `get_names()` helper function explicitly inside `plot_results()`, `map_results()` and `tabulate_results()` functions.
+
 # polAr 0.1.3 
 
 **CRAN release**